# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PeakNovel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  isPremium?: boolean;
  followerCount?: number;
  followingCount?: number;
  bookCount?: number;
  createdAt?: string;
}

export interface Book {
  id: number;
  title: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  genre: string;
  tags?: string[];
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  coverUrl: string | null;
  status: string;
  viewCount: number;
  chapterCount: number;
  rating: number;
  ratingCount?: number;
  isPremium?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  isBookmarked?: boolean | null;
  createdAt: string;
  updatedAt?: string;
}

export interface AuthorCard {
  id: number;
  username: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  followerCount: number;
  bookCount: number;
}

export interface HomeData {
  trending: Book[];
  topRanked: Book[];
  newReleases: Book[];
  featuredBooks: Book[];
  topAuthors: AuthorCard[];
  genres: string[];
}

export interface ChapterSummary {
  id: number;
  bookId: number;
  title: string;
  order: number;
  isPremium: boolean;
  wordCount?: number;
  createdAt: string;
}

export interface BookDetail {
  id: number;
  title: string;
  authorId: number;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  /** @nullable */
  authorBio?: string | null;
  authorFollowers?: number;
  genre: string;
  tags?: string[];
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  coverUrl: string | null;
  status: string;
  viewCount: number;
  chapterCount: number;
  rating: number;
  ratingCount?: number;
  isPremium?: boolean;
  isFeatured?: boolean;
  /** @nullable */
  isBookmarked?: boolean | null;
  /** @nullable */
  userRating?: number | null;
  /** @nullable */
  isFollowingAuthor?: boolean | null;
  chapters: ChapterSummary[];
  createdAt: string;
  updatedAt?: string;
}

export interface BookInput {
  title: string;
  genre: string;
  synopsis: string;
  coverUrl?: string;
  tags?: string[];
  status?: string;
  isPremium?: boolean;
}

export interface BookUpdate {
  title?: string;
  genre?: string;
  synopsis?: string;
  coverUrl?: string;
  tags?: string[];
  status?: string;
  isPremium?: boolean;
}

export interface BookList {
  books: Book[];
  total: number;
  page: number;
  totalPages: number;
}

export interface RatingInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  score: number;
}

export interface BookmarkStatus {
  bookmarked: boolean;
}

export interface Chapter {
  id: number;
  bookId: number;
  title: string;
  /** @nullable */
  content?: string | null;
  order: number;
  isPremium: boolean;
  wordCount?: number;
  createdAt: string;
  updatedAt?: string;
}

export interface ChapterRead {
  id: number;
  bookId: number;
  bookTitle: string;
  authorId?: number;
  authorName?: string;
  title: string;
  content: string;
  order: number;
  isPremium: boolean;
  wordCount?: number;
  /** @nullable */
  prevChapterId?: number | null;
  /** @nullable */
  nextChapterId?: number | null;
  createdAt?: string;
}

export interface ChapterInput {
  title: string;
  content: string;
  isPremium?: boolean;
  order?: number;
}

export interface ChapterUpdate {
  title?: string;
  content?: string;
  isPremium?: boolean;
  order?: number;
}

export interface Comment {
  id: number;
  chapterId: number;
  userId: number;
  username: string;
  /** @nullable */
  userAvatar?: string | null;
  content: string;
  likes: number;
  /** @nullable */
  isLiked?: boolean | null;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface UserProfile {
  id: number;
  username: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  role: string;
  isPremium?: boolean;
  followerCount: number;
  followingCount: number;
  bookCount: number;
  books?: Book[];
  /** @nullable */
  isFollowing?: boolean | null;
  createdAt: string;
}

export interface ProfileUpdate {
  username?: string;
  bio?: string;
  avatar?: string;
}

export interface FollowStatus {
  following: boolean;
}

export interface ReadingHistory {
  bookId: number;
  bookTitle: string;
  /** @nullable */
  coverUrl?: string | null;
  chapterId: number;
  chapterTitle: string;
  readAt: string;
}

export interface PremiumPlan {
  id: string;
  name: string;
  price: number;
  interval: string;
  features: string[];
}

export interface CheckoutInput {
  planId: string;
}

export interface CheckoutSession {
  url: string;
}

export interface SearchResults {
  books: Book[];
  authors: AuthorCard[];
  total: number;
}

export interface AdminStats {
  totalUsers: number;
  totalBooks: number;
  totalChapters: number;
  totalComments: number;
  premiumUsers: number;
  totalRevenue: number;
  newUsersToday?: number;
  newBooksToday?: number;
}

export interface AdminUser {
  id: number;
  username: string;
  email: string;
  role: string;
  isBanned: boolean;
  isPremium: boolean;
  bookCount?: number;
  createdAt: string;
}

export interface AdminUserList {
  users: AdminUser[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminUserUpdate {
  isBanned?: boolean;
  role?: string;
}

export interface AdminBookList {
  books: Book[];
  total: number;
  page: number;
  totalPages: number;
}

export interface AdminBookUpdate {
  isFeatured?: boolean;
  isHidden?: boolean;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  relatedBookId?: number | null;
  relatedChapterId?: number | null;
  actorName?: string | null;
  createdAt: string;
}

export interface NotificationList {
  notifications: Notification[];
  unreadCount: number;
}

export type SearchParams = {
q?: string;
genre?: string;
status?: string;
sort?: string;
page?: number;
};

export type ListBooksParams = {
genre?: string;
status?: string;
sort?: string;
authorId?: number;
page?: number;
};

export type ListAdminUsersParams = {
page?: number;
q?: string;
};

export type ListAdminBooksParams = {
page?: number;
q?: string;
};

export type MarkAllNotificationsRead200 = {
  ok?: boolean;
};