# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PeakNovel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── HOME ───────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Ban/unban or promote user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/books:
    get:
      operationId: listAdminBooks
      tags: [admin]
      summary: List all books for admin
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Books
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBookList"
  /admin/books/{id}:
    patch:
      operationId: updateAdminBook
      tags: [admin]
      summary: Feature/hide a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBookUpdate"
      responses:
        "200":
          description: Updated
    delete:
      operationId: deleteAdminBook
      tags: [admin]
      summary: Delete a book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ───────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        isPremium:
          type: boolean
        followerCount:
          type: integer
        followingCount:
          type: integer
        bookCount:
          type: integer
        createdAt:
          type: string

    # ── HOME SCHEMAS ──────────────────────────────────────
...
    Book:
      type: object
      required: [id, title, authorId, authorName, genre, status, coverUrl, viewCount, chapterCount, rating, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        genre:
          type: string
        tags:
          type: array
          items:
            type: string
        synopsis:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        status:
          type: string
        viewCount:
          type: integer
        chapterCount:
          type: integer
        rating:
          type: number
        ratingCount:
          type: integer
        isPremium:
          type: boolean
        isFeatured:
          type: boolean
        isBookmarked:
          type: ["boolean", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    AdminUserList:
      type: object
      required: [users, total, page, totalPages]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AdminUser:
      type: object
      required: [id, username, email, role, isBanned, isPremium, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        isBanned:
          type: boolean
        isPremium:
          type: boolean
        bookCount:
          type: integer
        createdAt:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        isBanned:
          type: boolean
        role:
          type: string
    AdminBookList:
      type: object
      required: [books, total, page, totalPages]
      properties:
        books:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
    AdminBookUpdate:
      type: object
      properties:
        isFeatured:
          type: boolean
        isHidden:
          type: boolean
...